Fix Bug 2123071 - add AES support for TMS server-side keygen on latest HSM / FIPS environment [RHCS 10.4].
This is the applet portion of this fix which consists of applet changes and TMS changes to be submitted separately.

This fix allows us to perform the AES_KEY_WRAP_KWP based unwrapping on the token for various key sizes. The reason why we
need KWP is because the latest HSM /FIPS environments only support KWP to be compatable between the hardware and software implementations of
AES_KEY_WRAP_PAD.

This fix is done by implementing KWP on top of AES_ECB, which is is assured to be present on various tokens supporting GP 2.2.x versions. Here we
also add a simple class to manipuate 64 bit unsigned quantities using only the "short" data type avaialable to us.

This fix has been tested up to importing a 3072 bit key. Testing revealed that such a key couldn't acutally to legal crypto operations to mininal code has been added to actually get that working.

This fix also contains some minor cleanup to remove unused code to not have the applet increase too much in final size.
This fix is so far only for the GP 2.2.1 branch, where we have assured support for AES on the tokens.

Include review suggestions.