Bug 2233158 - Make key wrapping algorithm configurable between AES-KWP and AES-CBC [RHCS 9.7.z].
This is the coolkey applet portion of the fix. Simply allow the caller to send a byte to choose between unwrapping
the private key with either the KWP or the CBC_PAD algorithm.

The incoming "keyType" byte from the server will observe the current 0x88 for KWP or 0x89 for CBC.
The CBC selection will result in a much quicker key unwrap operation since the support is intrinsic to the token that suppports
javacard api 3.04 and greater.

The Makefile has a couple of minor tweaks to compile against the newer api.

Modify cbc code to perform the decryption of the private key with CBC_NOPAD and perform the pkcs7 unpadding manually. This allows the code to work
without using the CBC with PKCS5 padding, which might not be supported on all tokens.